Fixed error with test-file5.md
test-file5.md was failure inducing input because it had text in between brackets and the parentheses for the link.  The resulting symptom was it took the "link" in the parentheses even though the brackets did not precede it.

The bug was that I did not check for preceding brackets.  An easy fix was to check the character preceding the parentheses; if it was not a closing bracket, terminate.